+++ Crossword/App/handlers/NavigateList.cs
