+++ Crossword/App/hint/QuickSolver.cs
