+++ Crossword/Crossword.csproj
@@ -66,7 +66,6 @@
     <PackageReference Include="FontStashSharp.MonoGame" Version="1.3.3" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Managers\" />
     <Folder Include="Parsers\" />
     <Folder Include="Screens\" />
     <Folder Include="UI\" />
